Fixed a bug that prevents a sync of changed users
If a user was changed by an second ldap import, before it was created, it's change state was updated by the LdapImportService from NEW to CHANGED.
This bug caused the AzureSyncService to try to update the user in Azure AD during the next sync instead of creating it. Since the user does not get an AzureImmutableId if it has not previously been created in Azure AD, the AzureSyncService tries to update the user with the AzureImmutableId "null".