Introduction of a new authentication system for the web module
The authentication system for the web module was completely rewritten. It is now possible to create individual users with username and password. For the access of applications to the API, api-keys can be created, which are sent to the backend in the X-API-Key header.
For each user or api-key it is possible to define if they get read-only or also write access.

See the API documentation for further and more detailed information.